+++ transport/transport.go
@@ -12,15 +12,8 @@ import (
 )
 
 var (
-	sep     = []byte("\n")
-	msgPool sync.Pool
+	sep = []byte("\n")
 )
-
-func init() {
-	msgPool.New = func() interface{} {
-		return &Message{}
-	}
-}
 
 const (
 	// MessageTypeKeepalive is the message type sent for keepalives--which are just an
@@ -132,8 +125,9 @@ type Transport interface {
 // WebSocket.
 type WebSocketTransport struct {
 	Connection *websocket.Conn
-	closed     bool
-	mutex      *sync.RWMutex
+	closed     bool // protected by writeMu
+	readMu     sync.Mutex
+	writeMu    sync.Mutex
 }
 
 // NewTransport creates an initialized Transport and return its pointer.
@@ -141,44 +135,45 @@ func NewTransport(conn *websocket.Conn) Transport {
 	return &WebSocketTransport{
 		Connection: conn,
 		closed:     false,
-		mutex:      &sync.RWMutex{},
 	}
 }
 
 // NewMessage creates a new Message.
 func NewMessage(msgType string, payload []byte) *Message {
-	msg := msgPool.Get().(*Message)
-	msg.Type = msgType
-	msg.Payload = payload
-	return msg
+	return &Message{
+		Type:    msgType,
+		Payload: payload,
+	}
 }
 
 // Close attempts to send a "going away" message over the websocket connection.
 // This will cause a Write over the websocket transport, which can cause a
 // panic. We rescue potential panics and consider the connection closed,
 // returning nil, because the connection _will_ be closed. Hay!
-func (t *WebSocketTransport) Close() error {
-	t.mutex.Lock()
-
-	defer func() {
-		// WriteMessage can annoyingly panic, because the websocket conn isn't safe
-		// for concurrent use. Recover here, and unlock the mutex.
-		_ = recover()
-		t.mutex.Unlock()
-	}()
+func (t *WebSocketTransport) Close() (err error) {
+	t.writeMu.Lock()
+	defer t.writeMu.Unlock()
 
 	if t.closed {
 		return nil
 	}
 
 	t.closed = true
+
+	defer func() {
+		cerr := t.Connection.Close()
+		if err == nil {
+			err = cerr
+		}
+	}()
+
 	return t.Connection.WriteMessage(websocket.CloseMessage, websocket.FormatCloseMessage(websocket.CloseGoingAway, "bye"))
 }
 
 // Closed returns true if the underlying websocket connection has been closed.
 func (t *WebSocketTransport) Closed() bool {
-	t.mutex.RLock()
-	defer t.mutex.RUnlock()
+	t.writeMu.Lock()
+	defer t.writeMu.Unlock()
 	return t.closed
 }
 
@@ -228,22 +223,19 @@ func (t *WebSocketTransport) Heartbeat(ctx context.Context, interval, timeout in
 // a ClosedError or a ConnectionError if unable to receive a message. Receive
 // blocks until the connection has a message ready or a timeout is reached.
 func (t *WebSocketTransport) Receive() (*Message, error) {
-	t.mutex.RLock()
-	if t.closed {
-		t.mutex.RUnlock()
+	t.readMu.Lock()
+	defer t.readMu.Unlock()
+
+	if t.Closed() {
 		return nil, ClosedError{"the websocket connection is no longer open"}
 	}
-	t.mutex.RUnlock()
 
 	_, p, err := t.Connection.ReadMessage()
 	if err != nil {
-		t.mutex.Lock()
-		t.closed = true
-		t.mutex.Unlock()
-
-		defer t.Connection.Close()
-
 		if websocket.IsCloseError(err, websocket.CloseGoingAway) {
+			t.writeMu.Lock()
+			defer t.writeMu.Unlock()
+			t.closed = true
 			return nil, ClosedError{err.Error()}
 		}
 		return nil, ConnectionError{err.Error()}
@@ -254,9 +246,7 @@ func (t *WebSocketTransport) Receive() (*Message, error) {
 		return nil, err
 	}
 
-	msg := msgPool.Get().(*Message)
-	msg.Type = msgType
-	msg.Payload = payload
+	msg := NewMessage(msgType, payload)
 	return msg, nil
 }
 
@@ -264,18 +254,17 @@ func (t *WebSocketTransport) Receive() (*Message, error) {
 // closed, returns a ClosedError. Returns a ConnectionError if the websocket
 // connection returns an error while sending, but the connection is still open.
 func (t *WebSocketTransport) Send(m *Message) (err error) {
-	defer msgPool.Put(m)
+	t.writeMu.Lock()
+	defer t.writeMu.Unlock()
+	if t.closed {
+		return ClosedError{"the websocket connection is no longer open"}
+	}
+
 	defer func() {
 		if m.SendCallback != nil {
 			m.SendCallback(err)
 		}
 	}()
-	t.mutex.RLock()
-	if t.closed {
-		t.mutex.RUnlock()
-		return ClosedError{"the websocket connection is no longer open"}
-	}
-	t.mutex.RUnlock()
 
 	msg := Encode(m.Type, m.Payload)
 	if err := t.Connection.WriteMessage(websocket.BinaryMessage, msg); err != nil {
@@ -283,9 +272,7 @@ func (t *WebSocketTransport) Send(m *Message) (err error) {
 		// because it's _really_ hard to figure out what errors from the
 		// websocket library are terminal and which aren't. So, abandon all
 		// hope, and reconnect if we get an error from the websocket lib.
-		t.mutex.Lock()
 		t.closed = true
-		t.mutex.Unlock()
 		if websocket.IsCloseError(err, websocket.CloseGoingAway) {
 			return ClosedError{err.Error()}
 		}
@@ -297,5 +284,7 @@ func (t *WebSocketTransport) Send(m *Message) (err error) {
 
 // SendCloseMessage sends a close control message over the transport
 func (t *WebSocketTransport) SendCloseMessage() (err error) {
+	t.writeMu.Lock()
+	defer t.writeMu.Unlock()
 	return t.Connection.WriteMessage(websocket.CloseMessage, websocket.FormatCloseMessage(websocket.CloseNormalClosure, ""))
 }